Update add-domain
Hello!
My name is Ivan, and I am seeking assistance. I am in Ukraine and have encountered a difficult situation. While searching for an online loan, I came across the website infozaem.com. There, I contacted a person who promised to help but requested an advance payment for their services. I made the payment, but after that, the person disappeared.

I reached out to the administrator of infozaem.com through the contact provided on https://infozaem.com/kontakti/ (mediagroup.feedback@gmail.com) and explained what had happened. However, instead of help, I was met with mockery and a statement that the site was created to scam people like me. I firmly declared that I would not let this matter go unnoticed.

I contacted the registrar hosting the site at abuse@namecheap.com to request assistance, but the registrar refused, citing a lack of evidence.

Even though it is sufficient that global antivirus programs identify infozaem.com as phishing, malicious, and malware, this was not enough for namecheap.com.

https://www.virustotal.com/gui/url/1962406b55c9568e1228e34df724197a4fbee78a5a6e757764de5b048a5bf9da

It is clear that the registrar does not want to lose its money since fraudulent sites like infozaem.com pay them. This means the registrar is breaking the law by hosting phishing sites on its platform.

I believe this website is fraudulent and involved in phishing, and I request that you block it so that others do not fall victim to scammers like those behind infozaem.com.